+++ src/collective/fingerpointing/utils.py
@@ -6,12 +6,10 @@ from zope.globalrequest import getRequest
 def get_request_information():
     """Return logged in user name and remote IP address."""
     request = getRequest()
-    user = api.user.get_current()
     try:
-        user_id = user.getMemberId()
+        user_id = api.user.get_current().getMemberId()
     except AttributeError:
-        # Anonymous user don't have getMemberId
-        user_id = user.getUserName()
+        user_id = '-'  # anonymous user
 
     # honor Cloudflare real client IP address request header if present
     # see: https://support.cloudflare.com/hc/en-us/articles/200170986
@@ -21,6 +19,6 @@ def get_request_information():
     elif 'HTTP_X_FORWARDED_FOR' in request.environ:
         ip = request.environ['HTTP_X_FORWARDED_FOR']
     else:
-        ip = request.getClientAddr() or 'None'
+        ip = request.getClientAddr() or 'None'  # return 'None' on tests
 
-    return user_id, ip  # returns ('test_user_1_', 'None') on tests
+    return user_id, ip